DTO Seguro y endpoints actualizados

+++ app/src/main/java/com/example/app_clinica_atl/data/remote/dto/SeguroDto.kt
@@ -2,12 +2,10 @@ package com.example.app_clinica_atl.data.remote.dto
 
 import com.google.gson.annotations.SerializedName
 
-/**
- * DTO que reemplaza a la entidad local `SeguroEntity`.
- */
+
 data class SeguroDto(
     @SerializedName("id") val id: Long = 0,
     @SerializedName("nombre") val name: String,
     @SerializedName("descripcion") val description: String,
-    @SerializedName("precio") val price: Double
+    @SerializedName("precio") val price: Int
 )
